Default settings on BrowseViewController are now externalized and are reloaded when the VC is reloaded between view transitions.  Should be easy now to just save them on the device across sessions.
The DemoWebService now retains data so that doing things like playing with the activation switch stick across cell reloads.

Activation is also wired to the web service now. WordPairCell's can now have their animation turned on and off via the BrowseViewModel.  This was important when navigating around.  Felt weird for them to animate because you were navigating the app but hadn't toggled the segmented control